Migrate classification service
This moves the classification module in ook.services and repackages it
as a service class. More work is likely needed to modernize this
service.
